+++ crates/buttplug_tests/tests/test_device_protocols.rs
@@ -150,6 +150,9 @@ async fn load_test_case(test_file: &str) -> DeviceTestCase {
 #[test_case("test_xuanhuan_protocol.yaml" ; "Xuanhuan Protocol")]
 #[test_case("test_yiciyuan_protocol.yaml" ; "Yiciyuan Protocol")]
 #[test_case("test_yiciyuan_protocol_fjb02.yaml" ; "Yiciyuan Protocol - FJB-02")]
+#[test_case("test_jejoue_protocol.yaml" ; "Je Joue Protocol")]
+#[test_case("test_maxpro_protocol.yaml" ; "Maxpro Protocol")]
+#[test_case("test_script_protocol.yaml" ; "Scripted Novel Protocol")]
 #[tokio::test]
 async fn test_device_protocols_embedded_v4(test_file: &str) {
   //tracing_subscriber::fmt::init();
@@ -282,6 +285,9 @@ async fn test_device_protocols_embedded_v4(test_file: &str) {
 #[test_case("test_xuanhuan_protocol.yaml" ; "Xuanhuan Protocol")]
 #[test_case("test_yiciyuan_protocol.yaml" ; "Yiciyuan Protocol")]
 #[test_case("test_yiciyuan_protocol_fjb02.yaml" ; "Yiciyuan Protocol - FJB-02")]
+#[test_case("test_jejoue_protocol.yaml" ; "Je Joue Protocol")]
+#[test_case("test_maxpro_protocol.yaml" ; "Maxpro Protocol")]
+#[test_case("test_script_protocol.yaml" ; "Scripted Novel Protocol")]
 #[tokio::test]
 async fn test_device_protocols_json_v4(test_file: &str) {
   //tracing_subscriber::fmt::init();
@@ -413,6 +419,9 @@ async fn test_device_protocols_json_v4(test_file: &str) {
 #[test_case("test_xuanhuan_protocol.yaml" ; "Xuanhuan Protocol")]
 #[test_case("test_yiciyuan_protocol.yaml" ; "Yiciyuan Protocol")]
 #[test_case("test_yiciyuan_protocol_fjb02.yaml" ; "Yiciyuan Protocol - FJB-02")]
+#[test_case("test_jejoue_protocol.yaml" ; "Je Joue Protocol")]
+#[test_case("test_maxpro_protocol.yaml" ; "Maxpro Protocol")]
+#[test_case("test_script_protocol.yaml" ; "Scripted Novel Protocol")]
 #[tokio::test]
 async fn test_device_protocols_embedded_v3(test_file: &str) {
   //tracing_subscriber::fmt::init();
@@ -545,6 +554,9 @@ async fn test_device_protocols_embedded_v3(test_file: &str) {
 #[test_case("test_xuanhuan_protocol.yaml" ; "Xuanhuan Protocol")]
 #[test_case("test_yiciyuan_protocol.yaml" ; "Yiciyuan Protocol")]
 #[test_case("test_yiciyuan_protocol_fjb02.yaml" ; "Yiciyuan Protocol - FJB-02")]
+#[test_case("test_jejoue_protocol.yaml" ; "Je Joue Protocol")]
+#[test_case("test_maxpro_protocol.yaml" ; "Maxpro Protocol")]
+#[test_case("test_script_protocol.yaml" ; "Scripted Novel Protocol")]
 #[tokio::test]
 async fn test_device_protocols_json_v3(test_file: &str) {
   //tracing_subscriber::fmt::init();
@@ -667,6 +679,9 @@ async fn test_device_protocols_json_v3(test_file: &str) {
 #[test_case("test_xuanhuan_protocol.yaml" ; "Xuanhuan Protocol")]
 #[test_case("test_yiciyuan_protocol.yaml" ; "Yiciyuan Protocol")]
 #[test_case("test_yiciyuan_protocol_fjb02.yaml" ; "Yiciyuan Protocol - FJB-02")]
+#[test_case("test_jejoue_protocol.yaml" ; "Je Joue Protocol")]
+#[test_case("test_maxpro_protocol.yaml" ; "Maxpro Protocol")]
+#[test_case("test_script_protocol.yaml" ; "Scripted Novel Protocol")]
 #[tokio::test]
 async fn test_device_protocols_embedded_v2(test_file: &str) {
   //tracing_subscriber::fmt::init();
@@ -790,6 +805,9 @@ async fn test_device_protocols_embedded_v2(test_file: &str) {
 #[test_case("test_xuanhuan_protocol.yaml" ; "Xuanhuan Protocol")]
 #[test_case("test_yiciyuan_protocol.yaml" ; "Yiciyuan Protocol")]
 #[test_case("test_yiciyuan_protocol_fjb02.yaml" ; "Yiciyuan Protocol - FJB-02")]
+#[test_case("test_jejoue_protocol.yaml" ; "Je Joue Protocol")]
+#[test_case("test_maxpro_protocol.yaml" ; "Maxpro Protocol")]
+#[test_case("test_script_protocol.yaml" ; "Scripted Novel Protocol")]
 #[tokio::test]
 async fn test_device_protocols_json_v2(test_file: &str) {
   util::device_test::client::client_v2::run_json_test_case(&load_test_case(test_file).await).await;
@@ -911,6 +929,9 @@ async fn test_device_protocols_json_v2(test_file: &str) {
 #[test_case("test_xuanhuan_protocol.yaml" ; "Xuanhuan Protocol")]
 #[test_case("test_yiciyuan_protocol.yaml" ; "Yiciyuan Protocol")]
 #[test_case("test_yiciyuan_protocol_fjb02.yaml" ; "Yiciyuan Protocol - FJB-02")]
+#[test_case("test_jejoue_protocol.yaml" ; "Je Joue Protocol")]
+#[test_case("test_maxpro_protocol.yaml" ; "Maxpro Protocol")]
+#[test_case("test_script_protocol.yaml" ; "Scripted Novel Protocol")]
 #[tokio::test]
 async fn test_device_protocols_embedded_v1(test_file: &str) {
   //tracing_subscriber::fmt::init();
@@ -1033,6 +1054,9 @@ async fn test_device_protocols_embedded_v1(test_file: &str) {
 #[test_case("test_xuanhuan_protocol.yaml" ; "Xuanhuan Protocol")]
 #[test_case("test_yiciyuan_protocol.yaml" ; "Yiciyuan Protocol")]
 #[test_case("test_yiciyuan_protocol_fjb02.yaml" ; "Yiciyuan Protocol - FJB-02")]
+#[test_case("test_jejoue_protocol.yaml" ; "Je Joue Protocol")]
+#[test_case("test_maxpro_protocol.yaml" ; "Maxpro Protocol")]
+#[test_case("test_script_protocol.yaml" ; "Scripted Novel Protocol")]
 #[tokio::test]
 async fn test_device_protocols_json_v1(test_file: &str) {
   util::device_test::client::client_v1::run_json_test_case(&load_test_case(test_file).await).await;